+++ src/main/java/org/tibennetwork/iamame/internetarchive/collectionitem/MessChdNeoCd.java
@@ -3,34 +3,18 @@ package org.tibennetwork.iamame.internetarchive.collectionitem;
 import java.io.File;
 import java.util.Set;
 
-import org.tibennetwork.iamame.mame.Software;
-
-public class MessChdNeoCd extends SoftwareListCollectionItem {
+public class MessChdNeoCd extends GenericChdCollectionItem {
 
     private String softwareFileUrlPattern 
         = "http://archive.org/download/MESS_0.149_CHD_neocd/" 
-            + "MESS_0.149_CHD_neocd.zip/MESS_0.149_CHD_neocd/%1$s/%2$s.chd";
+            + "MESS_0.149_CHD_neocd.zip/MESS_0.149_CHD_neocd/%1$s/%2$s";
 
     public MessChdNeoCd (Set<File> romsPaths, File writableRomPath) {
         super(romsPaths, writableRomPath);
     }
 
-    public void download (Software software) 
-            throws FileNotFoundInCollectionItem {
-        
-        String softwareName = software.getName();
-        String chdName = software.getChdName();
-        
-        String softwareFileUrl = String.format(
-            this.softwareFileUrlPattern,
-            softwareName,
-            chdName);
-
-        String destinationPath = this.writableRomPath.getAbsolutePath()
-            + File.separator
-            + software.getRelativeFilePath();
-
-        this.downloadFile(softwareFileUrl, destinationPath);
-
+    public String getSoftwareFileUrlPattern () {
+        return this.softwareFileUrlPattern;
     }
+
 }