+++ src/main/java/org/tibennetwork/iamame/internetarchive/collectionitem/MessChdSaturn.java
@@ -4,19 +4,20 @@ import java.io.File;
 import java.util.Set;
 
 import org.tibennetwork.iamame.mame.Software;
+import org.tibennetwork.iamame.mame.SoftwareFile;
 
 public class MessChdSaturn extends SoftwareListCollectionItem {
 
     private String[] softwareFileUrlPatterns = {
         "http://archive.org/download/MESS_0.149_CHD_saturn_1/" 
             + "MESS_0.149_CHD_saturn_1.zip/MESS_0.149_CHD_saturn_1/" 
-            + "%1$s/%2$s.chd",
+            + "%1$s/%2$s",
         "http://archive.org/download/MESS_0.149_CHD_saturn_2/" 
             + "MESS_0.149_CHD_saturn_2.zip/MESS_0.149_CHD_saturn_2/" 
-            + "%1$s/%2$s.chd",
+            + "%1$s/%2$s",
         "http://archive.org/download/MESS_0.149_CHD_saturn_3/" 
             + "MESS_0.149_CHD_saturn_3.zip/MESS_0.149_CHD_saturn_3/" 
-            + "%1$s/%2$s.chd"
+            + "%1$s/%2$s"
     };
 
     public MessChdSaturn (Set<File> romsPaths, File writableRomPath) {
@@ -26,30 +27,36 @@ public class MessChdSaturn extends SoftwareListCollectionItem {
     public void download (Software software) 
             throws FileNotFoundInCollectionItem {
         
-        String softwareName = software.getName();
-        String chdName = software.getChdName();
-       
-        for (String sfup: this.softwareFileUrlPatterns) {
-            String softwareFileUrl = String.format(
-                sfup,
-                softwareName,
-                chdName);
+        for (SoftwareFile sf: software.getMissingChdFiles(romsPaths)) {
 
             String destinationPath = this.writableRomPath.getAbsolutePath()
                 + File.separator
-                + software.getRelativeFilePath();
+                + sf.getRelativeFilePath();
 
-            try {
-                this.downloadFile(softwareFileUrl, destinationPath);
-                return;
-            } catch (FileNotFoundInCollectionItem e) {
-                // It's ok, it could be into another collection part
+            boolean fileFound = false;
+       
+            for (String sfup: this.softwareFileUrlPatterns) {
+
+                String softwareFileUrl = String.format(
+                    sfup,
+                    software.getName(),
+                    sf.getName());
+
+                try {
+                    this.downloadFile(softwareFileUrl, destinationPath);
+                    fileFound = true;
+                    break;
+                } catch (FileNotFoundInCollectionItem e) {
+                    // It's ok, it could be into another collection part
+                }
+
             }
-
+            
+            if (!fileFound) {
+                throw new FileNotFoundInCollectionItem(
+                    "The files have not been found on collection");
+            }
         }
-
-        throw new FileNotFoundInCollectionItem(
-            "The files have not been found on collection");
 
     }